dashboard_image/requirements.txt
bottle==0.12.19
Pillow==3.4.2
PyVirtualDisplay==3.0
selenium==3.141.0
Paste==3.5.0
